+++ src/spec/CompositeColliderSpec.ts
@@ -184,4 +184,27 @@ describe('A CompositeCollider', () => {
     expect(compCollider.contains(vec(-99.9, 0))).toBe(true);
     expect(compCollider.contains(vec(-101, 0))).toBe(false);
   });
+
+  it('is separated into a series of colliders in the dynamic tree', () => {
+    const compCollider = new ex.CompositeCollider([ex.Shape.Circle(50), ex.Shape.Box(200, 10, Vector.Half)]);
+
+    const dynamicTreeProcessor = new ex.DynamicTreeCollisionProcessor();
+    dynamicTreeProcessor.track(compCollider);
+
+    expect(dynamicTreeProcessor.getColliders().length).toBe(2);
+    expect(dynamicTreeProcessor.getColliders()[0] instanceof ex.CompositeCollider).toBe(false);
+    expect(dynamicTreeProcessor.getColliders()[1] instanceof ex.CompositeCollider).toBe(false);
+  });
+
+  it('removes all colliders in the dynamic tree', () => {
+    const compCollider = new ex.CompositeCollider([ex.Shape.Circle(50), ex.Shape.Box(200, 10, Vector.Half)]);
+
+    const dynamicTreeProcessor = new ex.DynamicTreeCollisionProcessor();
+    dynamicTreeProcessor.track(compCollider);
+
+    expect(dynamicTreeProcessor.getColliders().length).toBe(2);
+
+    dynamicTreeProcessor.untrack(compCollider);
+    expect(dynamicTreeProcessor.getColliders().length).toBe(0);
+  });
 });